Adds level to the Person record
Adds an integer level field to the Person record, defaulted to zero when not
provided. This field specifies the family tree's level the Person object is situated in.

Adds an additional copy method, withLevel, to create copy variants of Person with
different level value. Also, adds an incrementChildrenLevel method to
propagate an increment version of the parent level to its children.

Besides that, it adds specific comparison methods to ease the search process.