Enhances im- and mutability division in Person
When updating the children's generation, the Person's fields were
modified, which in some sense there is a mutability in this
operation. To define a better implementation between mutability
and immutability within this record class, incrementChildGeneration is
enhanced to generate a new Person record with its children's generation
incremented.

Also adds incrementChildGenerationIfAny to simplify readability and to
handle internally the increment in case the parent has any children.